fix using released pointer
In the case of multishot recv we can release operation only when there
will be no more completions for the same userdata; when
IOURING_CQE_F_MORE flag is not set.
If we release on first error there will be an cqe after that with
userdata > 0 but which don't point to the completion. Completion gone!